+++ src/main/scala/org/fs/mael/core/UserFriendlyException.scala
@@ -1,5 +1,10 @@
 package org.fs.mael.core
 
+/**
+ * An exception whose message alone is sufficient to let user know what's wrong.
+ *
+ * @author FS
+ */
 class UserFriendlyException(message: String)
   extends RuntimeException(message) {