+++ src/main/scala/org/fs/mael/ui/MainFrame.scala
@@ -4,6 +4,8 @@ import java.awt.Desktop
 import java.awt.Toolkit
 import java.awt.datatransfer.StringSelection
 
+import scala.util.Random
+
 import org.eclipse.jface.dialogs.MessageDialog
 import org.eclipse.jface.dialogs.MessageDialogWithToggle
 import org.eclipse.swt._
@@ -14,7 +16,9 @@ import org.fs.mael.BuildInfo
 import org.fs.mael.core.Status
 import org.fs.mael.core.backend.BackendManager
 import org.fs.mael.core.entry.DownloadEntryView
+import org.fs.mael.core.event.EventForUi
 import org.fs.mael.core.event.EventManager
+import org.fs.mael.core.event.Events._
 import org.fs.mael.core.event.UiSubscriber
 import org.fs.mael.core.list.DownloadListManager
 import org.fs.mael.core.utils.CoreUtils._
@@ -53,7 +57,7 @@ class MainFrame(
       layout.spacing = 0
     })
 
-    createMenu(shell)
+    val menu = new Menu(shell, SWT.BAR)
 
     val group = new Composite(shell, SWT.NONE)
     group.setLayout(new GridLayout().withCode { layout =>
@@ -63,17 +67,21 @@ class MainFrame(
       layout.marginHeight = 0
     })
 
-    createToolbar(group)
+    val toolbar = new ToolBar(group, SWT.FLAT)
 
     val sashForm = new SashForm(group, SWT.VERTICAL)
     sashForm.setLayoutData(new GridData(GridData.FILL, GridData.FILL, true, true))
 
-    createMainTable(sashForm)
-    createDetailsPanel(sashForm)
+    mainTable = new DownloadsTable(sashForm, resources)
+    logTable = new LogTable(sashForm, resources)
 
     sashForm.setWeights(Array(10, 10))
 
     // Init
+
+    fillMenu(menu)
+    fillToolbar(toolbar)
+    fillMainTable(mainTable)
 
     mainTable.init(downloadListMgr.list())
     mainTable.peer.setFocus()
@@ -86,11 +94,11 @@ class MainFrame(
     eventMgr.subscribe(subscriber)
   }
 
-  private def createMenu(parent: Decorations): Unit = {
-    val bar = new Menu(parent, SWT.BAR)
-    parent.setMenuBar(bar)
+  private def fillMenu(menu: Menu): Unit = {
+    val parent = menu.getParent
+    parent.setMenuBar(menu)
 
-    new MenuItem(bar, SWT.CASCADE).withCode { menuItem =>
+    new MenuItem(menu, SWT.CASCADE).withCode { menuItem =>
       menuItem.setText("&File")
 
       val submenu = new Menu(parent, SWT.DROP_DOWN)
@@ -101,7 +109,7 @@ class MainFrame(
       itemExit.addListener(SWT.Selection, tryExit)
     }
 
-    new MenuItem(bar, SWT.CASCADE).withCode { menuItem =>
+    new MenuItem(menu, SWT.CASCADE).withCode { menuItem =>
       menuItem.setText("&Service")
 
       val submenu = new Menu(parent, SWT.DROP_DOWN)
@@ -113,38 +121,35 @@ class MainFrame(
     }
   }
 
-  private def createToolbar(parent: Composite): Unit = {
-    val toolbar = new ToolBar(parent, SWT.FLAT)
-
+  private def fillToolbar(toolbar: ToolBar): Unit = {
     val btnAdd = (new ToolItem(toolbar, SWT.PUSH)).withCode { btnAdd =>
       btnAdd.setText("Add")
       btnAdd.addListener(SWT.Selection, e => {
-        val dialog = new Shell(shell, SWT.DIALOG_TRIM | SWT.APPLICATION_MODAL)
-        new AddDownloadFrame(dialog, resources, cfgMgr, backendMgr, downloadListMgr)
-        dialog.open()
+        val dialog = new EditDownloadDialog(None, shell, resources, cfgMgr, backendMgr, downloadListMgr, eventMgr)
+        dialog.peer.open()
       })
     }
 
     btnStart = (new ToolItem(toolbar, SWT.PUSH)).withCode { btnStart =>
       btnStart.setText("Start")
-      btnStart.setEnabled(false)
       btnStart.addListener(SWT.Selection, e => {
         mainTable.selectedEntries map { de =>
           val pair = backendMgr.getCastedPair(de)
           pair.backend.downloader.start(pair.de, cfgMgr.getProperty(ConfigOptions.NetworkTimeout))
         }
       })
+      btnStart.forDownloads(_ exists (_.status.canBeStarted))
     }
 
     btnStop = (new ToolItem(toolbar, SWT.PUSH)).withCode { btnStop =>
       btnStop.setText("Stop")
-      btnStop.setEnabled(false)
       btnStop.addListener(SWT.Selection, e => {
         mainTable.selectedEntries map { de =>
           val pair = backendMgr.getCastedPair(de)
           pair.backend.downloader.stop(pair.de)
         }
       })
+      btnStop.forDownloads(_ exists (_.status.canBeStopped))
     }
 
     toolbar.pack()
@@ -154,9 +159,7 @@ class MainFrame(
     })
   }
 
-  private def createMainTable(parent: Composite): Unit = {
-    mainTable = new DownloadsTable(parent, resources)
-
+  private def fillMainTable(mainTable: DownloadsTable): Unit = {
     val menu = new Menu(mainTable.peer).withCode { menu =>
       val parent = mainTable.peer
       parent.setMenu(menu)
@@ -175,19 +178,21 @@ class MainFrame(
         }
       }
 
+      val openProps = createMenuItem(menu, "Properties", parent, None) {
+        val deOption = mainTable.selectedEntryOption
+        require(deOption.isDefined)
+        val dialog = new EditDownloadDialog(deOption, shell, resources, cfgMgr, backendMgr, downloadListMgr, eventMgr)
+        dialog.peer.open()
+      }.forSingleDownloads()
+      mainTable.peer.addListener(SWT.MouseDoubleClick, e => openProps.notifyListeners(SWT.Selection, e))
+
       // TODO: Delete with file
       // TODO: Restart
-      // TODO: Properties
     }
 
     mainTable.peer.addListener(SWT.Selection, e => {
       logTable.render(mainTable.selectedEntryOption)
     })
-    mainTable.peer.addListener(SWT.Selection, e => updateButtonsEnabledState())
-  }
-
-  private def createDetailsPanel(parent: Composite): Unit = {
-    logTable = new LogTable(parent, resources)
   }
 
   private def onWindowClose(closeEvent: Event): Unit = {
@@ -294,12 +299,13 @@ class MainFrame(
     locations foreach Desktop.getDesktop.open
   }
 
-  private def updateButtonsEnabledState(): Unit = {
-    btnStart.setEnabled(mainTable.selectedEntries exists (_.status.canBeStarted))
-    btnStop.setEnabled(mainTable.selectedEntries exists (_.status.canBeStopped))
-  }
-
   private def clipboard = Toolkit.getDefaultToolkit.getSystemClipboard
+
+  /** Execute code in UI thread iff UI is not disposed yet */
+  private def syncExecSafely(code: => Unit): Unit =
+    if (!shell.isDisposed) display.syncExec { () =>
+      if (!shell.isDisposed) code
+    }
 
   //
   // Subscriber trait
@@ -311,27 +317,20 @@ class MainFrame(
     // TODO: Make per-download?
     val ProgressUpdateThresholdMs = 100
 
-    import org.fs.mael.core.event.EventForUi
-    import org.fs.mael.core.event.Events._
-
-    def fired(event: EventForUi): Unit = event match {
+    override def fired(event: EventForUi): Unit = event match {
       case Added(de) => syncExecSafely {
         if (!shell.isDisposed) {
           mainTable.add(de)
-          logTable.render(mainTable.selectedEntryOption)
-          updateButtonsEnabledState()
         }
       }
 
       case Removed(de) => syncExecSafely {
         mainTable.remove(de)
-        updateButtonsEnabledState()
       }
 
       case StatusChanged(de, prevStatus) => syncExecSafely {
         // Full row update
         mainTable.update(de)
-        updateButtonsEnabledState()
       }
 
       case Progress(de) =>
@@ -354,11 +353,28 @@ class MainFrame(
         }
       }
     }
+  }
 
-    /** Execute code in UI thread iff UI is not disposed yet */
-    private def syncExecSafely(code: => Unit): Unit =
-      if (!shell.isDisposed) display.syncExec { () =>
-        if (!shell.isDisposed) code
-      }
+  private implicit class ItemExt[T <: { def setEnabled(b: Boolean): Unit }](item: T) {
+    def forSingleDownloads(): T = {
+      mainTable.peer.addListener(SWT.Selection, e => {
+        item.setEnabled(mainTable.peer.getSelectionCount == 1)
+      })
+      item
+    }
+
+    def forDownloads(condition: Seq[DownloadEntryView] => Boolean): T = {
+      mainTable.peer.addListener(SWT.Selection, e => {
+        item.setEnabled(condition(mainTable.selectedEntries))
+      })
+      eventMgr.subscribe(new UiSubscriber {
+        override val subscriberId = "_event_forDownloads_" + Random.alphanumeric.take(20).mkString
+        override def fired(event: EventForUi): Unit = event match {
+          case StatusChanged(_, _) => syncExecSafely(item.setEnabled(condition(mainTable.selectedEntries)))
+          case _                   => // NOOP
+        }
+      })
+      item
+    }
   }
 }